+++ lib/mongory.rb
@@ -2,6 +2,7 @@
 
 require 'time'
 require 'date'
+require 'pp'
 require_relative 'mongory/version'
 require_relative 'mongory/utils'
 require_relative 'mongory/matchers'